Dev/django-with-react-rev3/requirements.txt
django~=3.0.0
djangorestframework~=3.11.0